requirements.txt
langchain==1.0.2
langchain-openai==1.0.1
langchain-community==0.4.1
langchain-ollama==0.3.10
python-dotenv==1.2.1

# Qwen model support (uses OpenAI-compatible API via Dashscope)
# No additional package needed - uses langchain-openai with custom base_url
pydantic>=2.0.0
fastapi==0.120.1
uvicorn==0.38.0
langchain-text-splitters>=0.3.0

# For document processing and embeddings
faiss-cpu==1.8.0
pypdf2==3.0.1
python-multipart==0.0.15

# For vector database memory
chromadb==0.4.22
langchain-chroma>=0.1.0

# For development
jupyter==1.1.1
streamlit==1.30.0

# Optional: for local models
# ollama==0.3.1